OLMIS-545 Add saving and getting user roles in user controller
Add code to get user roles and save user roles to user controller. A new DTO is necessary for these endpoints. Also since we are using Set for role assignments, and role assignments are value objects (they do not have their own identity), need to implement equals and hashCode for it.