Add a custom UUID generator
Some saves need to be done where a UUID is specified in the object to save. But since all saves us generated UUIDs, need to override the standard UUID generator to check for presence of UUID and use it to save, and do the standard thing if UUID is not specified. Change BaseEntity to use this new UUID generator.