OLMIS-545 Modify right and role controllers to streamline create/update
Right controller will not have a POST, as both create and update will be done by PUT. Role controller PUT does not need to check for existence, as it will save using the DTO in either case. The only difference is that UUID will be set (from URI) before persistence.