OLMIS-864, added link from OrderableProduct to ProgramProduct.
A number of issues surfaced relating to deserializing a ProgramProduct which should only have ID links to Programs, Products, etc.  And which have no identity outside of a database identity.  To help handle this a number of approaches were attempted, one didn't work because LoadTimeWeaving in SpringBoot isn't fully supported and adding AspectJ was non-trivial.  DTO's could have made made this much simpler, though also much more verbose and may actually be the solution that's chosen in the end as this weaves together domain entity life-cycle with serialization/de-serialization to a not so favorable degree.  One thing to document, is that @JsonDeserialize should never be mixed with a Jackson modified deserializer (here registered in Application.java) as it'll lead to a long debugging session to track down NullPointerExceptions.