OLMIS-1018 Re-enable requisition group program schedule controller integration tests
Fix some stuff in controller to pass tests; re-enable PUT endpoint. Fix equals and hashCode methods in domain object and DTO to use requisition group, program and processing schedule, as that is how a requisitionGroupProgramSchedule is defined by those properties.

For some reason, domain object did not have a reference to requisition group. Add that and fix repository integration test.