OLMIS-894 Split out JSON schema for right into separate file
Need to make sure copyRamlToBuild includes these JSON schema files to build directory. Remove copyRaml task, since it appears to be a duplicate of copyRamlToBuild. Using JSON schema draft 4, so remove refactor required keys into an array of property names. Draft 4 is used, as it is less verbose, and it seems to be a "default" of sorts (if rightArray does not specify a schema, draft 4 seems to be assumed).