OLMIS-1162 Update get supervised facilities endpoint to get by right and program
A user is mapped to supervised facilities by right and program.

Add a new validateId helper method to validate UUIDs. Add unit tests in the user object. Remove unit tests in the controller, because there is no real logic in the controller. Add new integration test for controller to handle bad request path.