OLMIS-1026 Allow searching users by extraData
Add extraData as another parameter by which to search. We need to use a separate search method in the repository, as JPA Criteria API does not have a way to search on Postgres JSON fields. We search the JSON field using a native SQL query, just passing in We merge the two search results and remove duplicates to get a final result.

Also add loginRestricted to search parameters since it seems to be missing.