OLMIS-1026 Change user search API to support extraData empty object
There seems to be some bug where when the reference data API is called with a null request body (extraData), the body inherits the request parameters for some reason. Address this by supporting an empty map for extraData. Change search method signatures to use a Map, to avoid PMD error of cyclomatic complexity. There also seem to be errors in using both the request parameters and the request body for searching with POST, so all search parameters are moved to the request body. Update RAML and tests.