OLMIS-1332 Update facility supported program model
Data model in Facility already has SupportedPrograms; add active and startDate fields to SupportedProgram (and add constructors). The startDate field needs to be a ZonedDateTime, but there is no pattern of saving and retrieving ZonedDateTime to/from the database, so create an attribute converter. It is auto-applied to all ZonedDateTime fields. The startDate has been annotated as a timestamp with time zone.

Fix serialization of supportedPrograms in Facility. FacilityDto was using ProgramDto for supportedPrograms; add new SupportedProgramDto (since we need to return extra fields now; active and startDate). Add exporter interface in SupportedProgram. SupportedProgramDto is different from SupportedProgram, as it does not have facility info, but does have program info and the two new fields.

Fix deserialization of supportedPrograms in Facility and update controller code. Do not fill in supportedProgram information when creating new Facility based on DTO, but do it in the controller (addSupportedProgramsToFacility). It does not make sense to use SupportedProgramDto to import, as it does not contain Facility information. The addSupportedProgramsToFacility handles getting all of the info necessary to construct the SupportedProgram object. Fix the bug in the PUT method which did not update supportedPrograms mapping. Add checks to return 400 Bad Request if supported program not found.

Add required active column data to demo data. Add controller integration tests because they did not exist for GETting all facilities, GETting one facility, POSTing facility and PUTting facility.