OLMIS-1456 Allow only service account to access save right endpoint
The PoC example is PUT /api/rights (saveRight), since only the account used by services should be able to use it. This account has the authority TRUSTED_CLIENT. ITs need to be refactored to mock two different calls to the auth service, one with a user-based token, and another with a client-based token (for services).