OLMIS-1456 Add code to show management of facilities
Add a FACILITIES_MANAGE right and add it to the Admin role. Add check to GET /facilities/id to allow access to trusted clients or user clients with the FACILITIES_MANAGE right. It can be tested by generating a service-level token and calling the endpoint, and also generating a user-level token and calling the endpoint, then removing the right from the Admin role and calling the endpoint again. It should work the first time and return a 403 Forbidden the second time.