OLMIS-1688 Fulfillment facilities are now filtered by right
The fulfillmentFacilities endpoint has been modified to take rightId
and uses is to filter the results by the right. Only those facilities
to which the user has got the right are returned. The tests have been
added/updated and RAML modified. Also the fulfillmentFacilities
and supervisedFacilities endpoints return FacilityDto instead
of raw Facility domain.