OLMIS-1605: Initialize JaVers repository and return 404 as appropriate
This check-in offers a workaround for <https://github.com/javers/javers/issues/214>. Specifically, upon application startup, it looks at every domain object associated with a JaversSpringDataAuditable repository. If any of them are absent from JaVers' repository, they are manually added.

This check-in also adds an integration test to ensure that a 404 error (rather than empty collection) is returned as necessary.

Finally, for development and testing, additional functionality has temporarily been added to UserController and UserRepositoryImpl. Some of it will later be removed.