OLMIS-1560 Add support for starter and demo data
Starter data SQL file would be present in db/starter. Demo data script should use afterMigrate.sql file name instead of data.sql because that's what Flyway uses to run SQL files after migration scripts.

Additionally add a conditional at the beginning of the demo data script to check for the existence of a table, which is created at the end of the script. This is so the script only runs once, the first time the service is deployed. These additional SQL statements are in the demo header and footer files.

Add a line to prepend the header and the starter data script, as well as append the footer to the demo data script. The starter data script is embedded into the demo data script because Flyway can only run one script with the name afterMigrate.sql.