OLMIS-1605: Change JaVers configuration
This commit updates JaVers' configuration such that it creates and uses tables within the "referencedata" rather than "public" schema.

It also adds @DiffIgnore to facility.supportedPrograms. This is in response to the problem mentioned at https://github.com/javers/javers/issues/234, and is an issue we've decided should be more thoroughly revisited after the 3.0 release.