Fix RG endpoints to allow associating facilities and schedules
There were two problems with the POST/PUT endpoints for the
RequisitionGroup. The endpoint was not accepting the
memberFacilities and schedules fields at all since the DTO
did not contain them. Another problem was incorrect bi-directional
association between RG and schedule while creating the RG via
an importer from DTO.