add spring boot health check
uses the most basic health endpoint provided by Spring Boot Actuator and then registers that HTTP endpoint with Consul.  The net effect for this level of effort is that Consul now has an actual mechanism to determine if the reference data service is functioning in slightly less clueless way.  The health endpoint can be expanded upon in the future to attach more meaning to the different http statuses returned to Consul.