OLMIS-2812, change user role assignments to eager fetch
Trialing an eager fetch on User's Role Assignments.  Why?  It's clear from the User.export method that GET /api/user/{id} always returns the full list of these Role Assignments.  As a OneToMany with a lazy fetch this results in the very undesirable N+1 select problem.  Since getting a user is used often (on the UI, in Auth, in checks from other services), increasing the speed of this endpoint, even marginally, may make more noticable differences elsewhere.  The downside to this change, is that this will need a considerably increased memory footprint.