OLMIS-2812, remove user from Auth principle
Loading the User object as the Auth principle seems unneded.  The only code that used this User object, really just pulled the User's UUID and then did another repository lookup.  Since this loading of the user in auth occurs even for endpoints that don't need to load the user, it feels like dead weight.  Instead the auth principle is now simple the User's UUID, and no find operation is executed.  Instead the lighter weight exists operation is used.