OLMIS-2813, upgrade to latest GA Spring Boot 1.5.4.
With the upgrade, two tests failed:  RequisitionGroupProgramScheduleIntegrationtest and JaversIntegrationTest.  The latter seemed more to be a pre-existing issue where the Javers test wasn't marked Transactional, and therefore sullied the DB.  The former feels highly problematic and unclear why it ever worked in the first place, and why it suddenly stopped working - nothing could be found in the Spring Boot 1.5.X release notes to indicate why.  The solution here is to be sure that there is a Transaction (as application code would have), that the transaction commits after the test, AND that the exception is thrown.  If all of these is true, then it appears the transaction does indeed fail, based on the deferred uniqueness constraint of RequisitionGroupProgramSchedule, and no tearDown method to wipe the database is needed.  That said, more investigation feels needed into deferred uniqueness constraints and testability.