OLMIS-2885 Associate right assignments to user directly
Associate right assignments to user directly, not through role assignments, since the role assignments are not important to right assignments. What is important is the user that is associated. When the user’s roles are updated, the right assignments are as well. However, the right assignments are not available through the user; they must be queried for through the RightAssignmentRepository.