OLMIS-3105 Remove post load supervisions to improve performance
Remove refreshing supervisions on PostLoad. This is causing the entire object graph to be eagerly fetched when getting a user. This also removes the home facility and supervised programs transient properties, and change /api/users/id/programs to get from the database. Also, adding role assignments is no longer in the User export, but should be done manually, to avoid eagerly fetching the entire object graph.

This code makes getting a user very fast, but the role assignments need to be returned with the user, so we need to add it back in.