OLMIS-3105 Add role assignment IDs to user for get user
Instead of having Hibernate get the role assignments, which eagerly fetches everything for some reason, manually get the role assignment IDs using a native query and a special resource with only the IDs. Had to use a named query and result set mapping because other approaches (projections) were having issues. This special resource is converted to a DTO and then returned. This keeps the fast performance while also returning the role assignment IDs.