OLMIS-3022 Call right assignment regen for requisition group update
Call the new right assignment service bean to regenerate right assignments on requisition group update. This requires for the right assignment generation code to know about the updated requisition group, so the save was changed to saveAndFlush. However, saveAndFlush is only present in the JpaRepository, which extends PagingAndSortingRepository. That is why RequisitionGroupRepository now extends JpaRepository.

Asynchronous running of right assignments also added, by annotating the regenerateRightAssignments with Async. Also, the Application class needs EnableAsync for it to work.