OLMIS-3022 Change APIs for regen to flush after save
All APIs that need to regenerate right assignments need to immediately flush after saving, so that the regenerate code can pick up the changes. To be able to call saveAndFlush, needed to refactor repositories to extend JpaRepository instead of PagingAndSortingRepository. Some tests were affected.

For some reason, FacilityRepository was overriding some save methods that were exactly the same as the methods being overridden.