OLMIS-4017 Create separate types of dispensables
Dispensable concept is now separated out into types. There are three types for now—default, container, and vaccine. Container is for things that contain other things; currently only used for syringes of different sizes. Vaccine is for vaccines, for number of doses and the route of administration into the body. Default is for everything else.

The dispensable DTO also returns displayUnit, which is the value from toString.