OLMIS-4800, remove agent from pipeline and give to work-stages
By removing the agent from the pipeline we should allow the declaritive pipeline pieces to use the lightweight executor (i.e. one that doesn't take up a real one), and then we need to procure an agent in stages that do real work.  Declaring an agent for stages that just call external jobs is avoided as those external jobs will assign their own executor.  By using individual agents in each stage, it means we'll need to stash/unstash the source code as we go.  We may also need to do the same with the docker image, however first we'll try without.